# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/v1/spells/": {
            "get": {
                "description": "get spells by filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spells"
                ],
                "summary": "Get spell list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.prettySpell"
                            }
                        }
                    }
                }
            }
        },
        "/v1/url-sets/": {
            "post": {
                "description": "Create empty url set without name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url-sets"
                ],
                "summary": "Create url set",
                "responses": {
                    "200": {
                        "description": "link to url set",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/url-sets/{unique}": {
            "get": {
                "description": "Get basic info about url set. Just name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url-sets"
                ],
                "summary": "Get url set",
                "parameters": [
                    {
                        "type": "string",
                        "description": "url set unique link part",
                        "name": "unique",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.prettySetDto"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url-sets"
                ],
                "summary": "Rename url set",
                "parameters": [
                    {
                        "type": "string",
                        "description": "url set unique link part",
                        "name": "unique",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updating data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.renameUrlSetDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/url-sets/{unique}/add/{spellId}": {
            "post": {
                "description": "Can add spell to url set",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url-sets"
                ],
                "summary": "Add spell to url set",
                "parameters": [
                    {
                        "type": "string",
                        "description": "url set unique link part",
                        "name": "unique",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of spell",
                        "name": "spellId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/url-sets/{unique}/all-spells": {
            "get": {
                "description": "get all spells. Overview for adding/removing spells to/from set",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url-sets"
                ],
                "summary": "Get list of all spells",
                "parameters": [
                    {
                        "type": "string",
                        "description": "url set unique link part",
                        "name": "unique",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.prettySpell"
                            }
                        }
                    }
                }
            }
        },
        "/v1/url-sets/{unique}/spells": {
            "get": {
                "description": "Get only spells in url set",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url-sets"
                ],
                "summary": "Get list of url set spells",
                "parameters": [
                    {
                        "type": "string",
                        "description": "url set unique link part",
                        "name": "unique",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.prettySpell"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "v1.prettySetDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
        "v1.prettySpell": {
            "type": "object",
            "properties": {
                "casting_time": {
                    "type": "string"
                },
                "classes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "distance": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "has_material_component": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_ritual": {
                    "type": "boolean"
                },
                "is_somatic": {
                    "type": "boolean"
                },
                "is_verbal": {
                    "type": "boolean"
                },
                "level": {
                    "type": "integer"
                },
                "magical_school": {
                    "type": "string"
                },
                "material_component": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                },
                "source_name": {
                    "type": "string"
                }
            }
        },
        "v1.renameUrlSetDto": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}